week4/requirements.txt
PyMuPDF==1.23.14
sentence-transformers==2.2.2
faiss-cpu==1.7.4
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.24.3
pandas==2.0.3
tiktoken==0.5.1
jupyter==1.0.0
requests==2.31.0
python-multipart==0.0.6